Moves Signup/Login to components, Creates Single Portal Page
